Refactory some Python code into a proper package
