+++ CMakeLists.txt
@@ -12,6 +12,7 @@ endif()
 
 # find dependencies
 find_package(ament_cmake REQUIRED)
+find_package(ament_cmake_python REQUIRED)
 find_package(rclcpp REQUIRED)
 
 add_executable(node_spinning src/node_spinning.cpp)
@@ -43,6 +44,8 @@ if(ENABLE_SYSTEM_METRIC_COLLECTOR)
     EXPORT export_${PROJECT_NAME}
     DESTINATION lib/${PROJECT_NAME})
 endif()
+
+ament_python_install_package(${PROJECT_NAME})
 
 install(DIRECTORY templates
   DESTINATION share/${PROJECT_NAME})
