+++ test/test_spinning.py.in
@@ -7,15 +7,14 @@ import sys
 import tempfile
 import unittest
 
+from buildfarm_perf_tests.launch import attach_system_metric_collector
+from buildfarm_perf_tests.test_results import read_performance_test_csv
+from buildfarm_perf_tests.test_results import write_jenkins_plot_csv
 from launch import LaunchDescription
-from launch.substitutions import LaunchConfiguration, EnvironmentVariable
-import ament_index_python
+from launch.actions import OpaqueFunction
+from launch.substitutions import LaunchConfiguration
 from launch_ros.actions import Node
 import launch_testing
-from launch.actions import (
-    ExecuteProcess, OpaqueFunction,
-    RegisterEventHandler, SetEnvironmentVariable)
-from launch.event_handlers import OnProcessStart
 import matplotlib  # noqa: F401
 import matplotlib.pyplot as plt
 import numpy as np  # noqa: F401
@@ -29,9 +28,9 @@ def _cleanUpLogs(node_spinning_log_prefix):
         os.remove(log)
 
 
-def _raw_to_csv(csv_data, csv_path):
+def _raw_to_csv(dataframe, csv_path):
     """
-    Convert from the raw csv format to the csv for the Jenkins plot plugin.
+    Convert from the raw csv data to the csv for the Jenkins plot plugin.
 
     Do not change the order of the columns. The plot plugin indexes into the
     csv using the column number instead of the column name, because we're
@@ -40,8 +39,6 @@ def _raw_to_csv(csv_data, csv_path):
     Changing the column names here will change the name of the line that
     appears on the plot.
     """
-    dataframe = pd.read_csv(csv_data, skiprows=0, sep='[ \t]*,[ \t]*',
-                            engine='python')
     dataframe_agg = dataframe.agg(['median', 'mean'])
 
     values = [
@@ -59,14 +56,10 @@ def _raw_to_csv(csv_data, csv_path):
       str(dataframe['resident anonymous memory (Mb)'].describe(percentiles=[0.95]).iloc[5]),
     ]
 
-    with open(csv_path, 'w') as csv:
-        csv.write(','.join(['@TEST_NAME@'] * len(values)) + '\n')
-        csv.write(','.join(values) + '\n')
+    write_jenkins_plot_csv(csv_path, '@TEST_NAME@', values)
 
 
-def _raw_to_png(csv_data, png_path):
-    dataframe = pd.read_csv(csv_data, skiprows=0, sep='[ \t]*,[ \t]*',
-                            engine='python')
+def _raw_to_png(dataframe, png_path):
     pd.options.display.float_format = '{:.4f}'.format
     dataframe[['T_experiment', 'virtual memory (Mb)']].plot(x='T_experiment')
 
@@ -90,27 +83,9 @@ def _raw_to_png(csv_data, png_path):
     plt.savefig(png_path[:-4] + "_resident_anonymous_memory.png")
 
 
-def on_process_start(process_started, launch_context, observation_process):
-    SetEnvironmentVariable(
-        'PROCESS_UNDER_TEST_ID', str(process_started.pid)).execute(launch_context)
-    observation_process.execute(launch_context)
-
 def generate_test_description(ready_fn):
     node_spinning_log_prefix = tempfile.mkstemp(
         prefix='overhead_node_test_results_@TEST_NAME@_', text=True)[1]
-
-    node_metrics_collector = ExecuteProcess(
-      cmd=[
-          os.path.join(
-              ament_index_python.get_package_prefix('buildfarm_perf_tests'),
-              'lib/buildfarm_perf_tests',
-              'system_metric_collector',
-          ),
-          # Arguments
-          "--timeout", "@NODE_SPINNING_TIMEOUT@",
-          "--log", node_spinning_log_prefix,
-          "--process_pid", EnvironmentVariable('PROCESS_UNDER_TEST_ID')
-      ],)
 
     node_spinning_test = Node(
         package='buildfarm_perf_tests',
@@ -119,12 +94,13 @@ def generate_test_description(ready_fn):
         sigterm_timeout=LaunchConfiguration('sigterm_timeout', default=@NODE_SPINNING_TIMEOUT@+2)
     )
 
+    node_metrics_collector, node_spinning_test_hook = attach_system_metric_collector(
+        node_spinning_test, node_spinning_log_prefix, timeout=@NODE_SPINNING_TIMEOUT@)
+
     return LaunchDescription([
-      RegisterEventHandler(OnProcessStart(
-          target_action=node_spinning_test,
-          on_start=lambda p, c: on_process_start(p, c, node_metrics_collector)
-        )),
+        node_spinning_test_hook,
         node_spinning_test,
+        launch_testing.util.KeepAliveProc(),
         OpaqueFunction(function=lambda context: ready_fn()),
     ]), locals()
 
@@ -152,16 +128,18 @@ class NodeSpinningTestResults(unittest.TestCase):
         node_spinning_logs = glob(node_spinning_log_prefix + '*')
 
         if node_spinning_logs:
+            node_spinning_data = read_performance_test_csv(node_spinning_logs[0], start_marker=None)
+
             node_spinning_report_csv = os.environ.get('PERFORMANCE_OVERHEAD_NODE_CSV')
             if node_spinning_report_csv:
-                _raw_to_csv(node_spinning_logs[0], node_spinning_report_csv)
+                _raw_to_csv(node_spinning_data, node_spinning_report_csv)
             else:
                 print('No CSV report written - set PERFORMANCE_OVERHEAD_NODE_CSV to write a report',
                       file=sys.stderr)
 
             node_spinning_report_png = os.environ.get('PERFORMANCE_OVERHEAD_NODE_PNG')
             if node_spinning_report_png:
-                _raw_to_png(node_spinning_logs[0], node_spinning_report_png)
+                _raw_to_png(node_spinning_data, node_spinning_report_png)
             else:
                 print('No PNG report written - set PERFORMANCE_OVERHEAD_NODE_PNG to write a report',
                       file=sys.stderr)
